[types] Move types back to Typography file.
